+++ CHANGELOG.md
@@ -33,6 +33,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2159 Get rid of connectWithTransitionGroup in viz
 - YLP-2162 Migrate CbgDateTrace store from redux to TrendsProvider
 - YLP-2176 Migrate CBG slice actions from redux to TrendsProvider
+- YLP-2178 Remove redux store
 
 ### Fixed
 - YLP-2145 Patient cannot see monitoring widgets when he's in a monitoring team and another non-monitoring team